by-covid-xml-transformer/requirements.txt
# Direct dependencies
ConfigArgParse==1.5.3
lxml==4.9.1
prettierfier==1.0.3
py12flogging==0.5.0
pyoai==2.5.0
PyYAML==6.0
requests==2.28.0
saxonche==12.3.0
wheel==0.43.0


# Indirect dependencies
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
six==1.16.0
urllib3==1.26.12
